@@ -1,5 +1,10 @@
 module github.com/lightstep/varopt
 
-go 1.15
+go 1.21
 
-require github.com/stretchr/testify v1.4.0
+require (
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/rogpeppe/go-internal v1.10.0 // indirect
+	github.com/stretchr/testify v1.8.4
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+)
